fix(base-driver): use new mixin strategy
also fix a broken type in fake-driver
