--- packages/base-driver/lib/basedriver/commands/execute.js
@@ -1,50 +1,0 @@
-import _ from 'lodash';
-import {errors, validateExecuteMethodParams} from '../../protocol';
-
-/**
- * @template {Constraints} C
- * @param {import('./session').SessionBase<C>} Base
- * @returns {ExecuteBase<C>}
- */
-export function ExecuteMixin(Base) {
-  /**
-   * @implements {IExecuteCommands}
-   */
-  class ExecuteCommands extends Base {
-    /**
-     * @param {string} script
-     * @param {[Record<string, any>]|[]} protoArgs
-     */
-    async executeMethod(script, protoArgs) {
-      const Driver = /** @type {import('@appium/types').DriverClass<Driver>} */ (this.constructor);
-      const commandMetadata = {...Driver.executeMethodMap?.[script]};
-      if (!commandMetadata.command) {
-        const availableScripts = _.keys(Driver.executeMethodMap);
-        throw new errors.UnsupportedOperationError(
-          `Unsupported execute method '${script}'. Available methods ` +
-            `are: ${availableScripts.join(', ')}`
-        );
-      }
-      const command = /** @type {import('@appium/types').DriverCommand} */ (
-        this[commandMetadata.command]
-      );
-      const args = validateExecuteMethodParams(protoArgs, commandMetadata.params);
-      return await command.call(this, ...args);
-    }
-  }
-  return ExecuteCommands;
-}
-
-/**
- * @typedef {import('@appium/types').IExecuteCommands} IExecuteCommands
- * @typedef {import('@appium/types').Driver} Driver
- * @typedef {import('@appium/types').Constraints} Constraints
- */
-
-/**
- * @template {Constraints} C
- * @typedef {import('../driver').BaseDriverBase<C, import('@appium/types').ITimeoutCommands &
- * import('@appium/types').IEventCommands & import('@appium/types').IFindCommands &
- * import('@appium/types').ILogCommands & import('@appium/types').ISettingsCommands &
- * import('@appium/types').ISessionCommands & IExecuteCommands>} ExecuteBase
- */
