+++ packages/base-driver/lib/basedriver/commands/session.ts
@@ -1,0 +1,35 @@
+import {Constraints, ISessionCommands, MultiSessionData} from '@appium/types';
+import _ from 'lodash';
+import {BaseDriver} from '../driver';
+import {mixin} from './mixin';
+
+declare module '../driver' {
+  interface BaseDriver<C extends Constraints> extends ISessionCommands {}
+}
+
+const SessionCommands: ISessionCommands = {
+  async getSessions<C extends Constraints>(this: BaseDriver<C>) {
+    let ret: MultiSessionData[] = [];
+
+    if (this.sessionId) {
+      ret.push({
+        id: this.sessionId,
+        capabilities: this.caps,
+      });
+    }
+
+    return ret;
+  },
+
+  /**
+   * Returns capabilities for the session and event history (if applicable)
+   */
+  async getSession<C extends Constraints>(this: BaseDriver<C>) {
+    if (this.caps.eventTimings) {
+      return {...this.caps, events: this.eventHistory};
+    }
+    return this.caps;
+  },
+};
+
+mixin(SessionCommands);
